requirements.txt
pillow >= 8.0
qrcode >= 7.0
